feat: allow monitoring image to collect gpu metrics
in order to support using pynvml without blowing up the size
of the monitoring docker image, tasks need to run on the non-default
batch-debian vm image. This makes google batch set up the nvidia
drivers on the vm and mount them to the container instead of
relying on the drivers being installed on the monitor image.

An added benefit is reduced docker image sizes for all tasks
that require a gpu.